Add feature specs for chat online buttons
Add feature spec coverage for all the chat online buttons/variations. This was
challenging because we need to be able to mock the current browser time from
Rails for Capybara, which is in a different process. The method I used involves
setting the time you want to use in the query string and then loading in a new
pack that mocks out the current Javascript `Date` object. I used `sinon` for
this instead of `jest` becase you can't include `jest` outside of the test
environment. The downside is we now have a test dependency in our non-dev JS
dependencies, but its only loaded in on the test environment so it won't have a
performance hit.